requirements.txt
attrs==19.3.0
autopep8==1.5
backcall==0.1.0
certifi==2020.4.5.1
chardet==3.0.4
Click==7.0
cm==0.1a2
colorama==0.4.1
crypto==1.4.1
decorator==4.4.1
stripe>=2.48.0,<3.0
defusedxml==0.6.0
Django==2.2.5
django-allauth==0.41.0
django-appconf==1.0.3
django-autocomplete-light==3.5.1
django-betterforms==1.2
django-ckeditor==5.7.1
django-cors-headers==3.3.0
django-debug-toolbar==2.2
django-filter==2.2.0
django-formtools==2.1
django-js-asset==1.2.2
django-rest-framework==0.1.0
django-select2==7.2.0
django-widget-tweaks==1.4.5
djangorestframework==3.11.0
drf-url-filters==0.5.1
et-xmlfile==1.0.1
html5lib==1.0.1
httplib2==0.17.2
idna==2.8
ipdb==0.12.2
ipython==7.9.0
ipython-genutils==0.2.0
jdcal==1.4.1
jedi==0.15.1
Naked==0.1.31
oauthlib==3.1.0
openpyxl==3.0.3
paragraph==1.1.0
parso==0.5.1
pickleshare==0.7.5
Pillow==6.2.1
prompt-toolkit==2.0.10
psycopg2==2.8.5
pycodestyle==2.5.0
Pygments==2.4.2
pylint==2.4.3
PyPDF2==1.26.0
python3-openid==3.1.0
pytz==2019.2
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-oauthlib==1.3.0
shellescape==3.8.1
six==1.13.0
sqlparse==0.3.0
traitlets==4.3.3
urllib3==1.25.8
views==0.3
virtualenv==16.7.5
voluptuous==0.11.7
wcwidth==0.1.7
webencodings==0.5.1
xhtml2pdf==0.2.2
xlwt==1.3.0
asdad
